requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
langchain==0.3.0 
langchain-google-genai==2.0.0 
langchain-community==0.3.0 
pypdf2==3.0.1 
beautifulsoup4==4.12.3 
chromadb==0.5.5
python-dotenv==1.0.1
python-multipart==0.0.20